Fixed expire time in thrift msg to int64.

+++ source/fdsp/fds_service.thrift
@@ -1082,7 +1082,7 @@ struct CtrlObjectMetaDataPropagate
     10: i32              objectFlags
     
     /* object expieration time */
-    11: i32              objectExpireTime    
+    11: i64              objectExpireTime
 }
 
 struct CtrlObjectRebalanceDeltaSet
